+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Extend chart component options ([PR #4424](https://github.com/alphagov/govuk_publishing_components/pull/4424))
+* Use the first organisation to generate breadcrumbs, if no parent linked ([PR #4270](https://github.com/alphagov/govuk_publishing_components/pull/4270))
 
 ## 45.6.1